+++ dao/src/main/java/com/mtons/mblog/pojo/ChatUserRecordExample.java
@@ -315,52 +315,52 @@ public class ChatUserRecordExample {
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdEqualTo(Integer value) {
+        public Criteria andChatGroupIdEqualTo(Long value) {
             addCriterion("chat_group_id =", value, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdNotEqualTo(Integer value) {
+        public Criteria andChatGroupIdNotEqualTo(Long value) {
             addCriterion("chat_group_id <>", value, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdGreaterThan(Integer value) {
+        public Criteria andChatGroupIdGreaterThan(Long value) {
             addCriterion("chat_group_id >", value, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andChatGroupIdGreaterThanOrEqualTo(Long value) {
             addCriterion("chat_group_id >=", value, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdLessThan(Integer value) {
+        public Criteria andChatGroupIdLessThan(Long value) {
             addCriterion("chat_group_id <", value, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdLessThanOrEqualTo(Integer value) {
+        public Criteria andChatGroupIdLessThanOrEqualTo(Long value) {
             addCriterion("chat_group_id <=", value, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdIn(List<Integer> values) {
+        public Criteria andChatGroupIdIn(List<Long> values) {
             addCriterion("chat_group_id in", values, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdNotIn(List<Integer> values) {
+        public Criteria andChatGroupIdNotIn(List<Long> values) {
             addCriterion("chat_group_id not in", values, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdBetween(Integer value1, Integer value2) {
+        public Criteria andChatGroupIdBetween(Long value1, Long value2) {
             addCriterion("chat_group_id between", value1, value2, "chatGroupId");
             return (Criteria) this;
         }
 
-        public Criteria andChatGroupIdNotBetween(Integer value1, Integer value2) {
+        public Criteria andChatGroupIdNotBetween(Long value1, Long value2) {
             addCriterion("chat_group_id not between", value1, value2, "chatGroupId");
             return (Criteria) this;
         }
